third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/hostcontext-first-child-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Test that :host-context(:first-child) is re-evaluated when :first-child changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
